Minor fixes.
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@4668 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/main/java/org/red5/server/stream/PlaylistSubscriberStream.java
@@ -280,15 +280,17 @@ public class PlaylistSubscriberStream extends AbstractClientStream implements IP
 
 	/** {@inheritDoc} */
 	public void close() {
-		engine.close();
-		onChange(StreamState.CLOSED);
-		items.clear();
-		// clear jobs
-		if (schedulingService != null && !jobs.isEmpty()) {
-			for (String jobName : jobs) {
-				schedulingService.removeScheduledJob(jobName);
-			}
-			jobs.clear();
+		if (engine != null) {
+    		engine.close();
+    		onChange(StreamState.CLOSED);
+    		items.clear();
+    		// clear jobs
+    		if (schedulingService != null && !jobs.isEmpty()) {
+    			for (String jobName : jobs) {
+    				schedulingService.removeScheduledJob(jobName);
+    			}
+    			jobs.clear();
+    		}
 		}
 	}
